requirements.txt
fastapi[standard]==0.112.0,<0.113.0
pydantic>=2.7.0,<3.0.0
pydantic-core<=2.33.0
psutil
py-cpuinfo
gputil
